Extract Trade Abuse checking from individual tradebots
